test: testing reservation form (Failing tests)
These test currently does pass as intended due to bad XHR post request
when submitting the reservation form data to Firebase Firestore. Tried
changing the firestore settings as per the guide on this open issue:
https://github.com/cypress-io/cypress/issues/2374
which does not work as it only allow Cypress to make one post request
on the first testm, any sebsequent test that needs to post data
to firestore will fail with bad XHR post request.